src/main/java/application/controller/dto/BetDTO.java
package application.controller.dto;

import java.io.Serializable;

public class BetDTO implements Serializable {

    private String name;
    private String accountId;
    private Double stake;

    public Double getStake() {
        return stake;
    }

    public void setStake(Double stake) {
        this.stake = stake;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getAccountId() {
        return accountId;
    }

    public void setAccountId(String accountId) {
        this.accountId = accountId;
    }

}
